requirements.txt
fastapi==0.95.2
sqlalchemy==2.0.4
pymysql==1.0.2
mangum==0.13.0
cloud-sql-python-connector==1.13.0
google-cloud-storage==2.18.2
uvicorn==0.32.0
numpy==2.1.3
pandas==2.2.3
pyyaml==6.0.2
